feat: 企画一覧で直接一覧を表示するようにした
