+++ src/components/project/ThumbnailImage.tsx
